adding spirals separating out the persteps
